requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-generativeai==0.8.0
requests==2.32.3
pydantic==2.8.2
python-dotenv==1.0.0
